Commented out auth guards temporarily (for development communication … (#1510)
* Commented out auth guards temporarily (for development communication tests). Validations added and comments

* Validation added

* Added @IsNotEmpty() to data in grant dto

* Moved granttypes ctrl to admin api. Get by PK added. Obsolete id keys rmoved

+++ libs/auth-api-lib/src/lib/services/clients.service.ts
@@ -1,7 +1,6 @@
-import { Inject, Injectable } from '@nestjs/common'
+import { BadRequestException, Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
-import { Sequelize } from 'sequelize-typescript'
 import { Client } from '../entities/models/client.model'
 import { ClientAllowedScope } from '../entities/models/client-allowed-scope.model'
 import { ClientAllowedCorsOrigin } from '../entities/models/client-allowed-cors-origin.model'
@@ -16,21 +15,22 @@ import { ClientClaim } from '../entities/models/client-claim.model'
 
 @Injectable()
 export class ClientsService {
-  private clients: Client[]
-
   constructor(
-    private sequelize: Sequelize,
     @InjectModel(Client)
     private clientModel: typeof Client,
     @Inject(LOGGER_PROVIDER)
     private logger: Logger,
   ) {}
 
+  /** Gets a client by it's id */
   async findClientById(id: string): Promise<Client> {
     this.logger.debug(`Finding client for id - "${id}"`)
 
-    return this.clientModel.findOne({
-      where: { clientId: id },
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
+
+    return this.clientModel.findByPk(id, {
       include: [
         ClientAllowedScope,
         ClientAllowedCorsOrigin,
@@ -45,14 +45,20 @@ export class ClientsService {
     })
   }
 
+  /** Creates a new client */
   async create(client: ClientDTO): Promise<Client> {
     this.logger.debug('Creating a new client')
 
     return await this.clientModel.create({ ...client })
   }
 
+  /** Updates an existing client */
   async update(client: ClientUpdateDTO, id: string): Promise<Client> {
     this.logger.debug('Updating client with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('id must be provided')
+    }
 
     await this.clientModel.update(
       { ...client },
@@ -64,8 +70,13 @@ export class ClientsService {
     return await this.findClientById(id)
   }
 
+  /** Deletes a client by id */
   async delete(id: string): Promise<number> {
     this.logger.debug('Deleting client with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('id must be provided')
+    }
 
     return await this.clientModel.destroy({
       where: { clientId: id },
